split hist maker to smaller files
