+++ plugin/CMakeLists.txt
@@ -2,16 +2,10 @@ if (PLUGIN_DENSE_PARSER)
   target_sources(objxgboost PRIVATE ${xgboost_SOURCE_DIR}/plugin/dense_parser/dense_libsvm.cc)
 endif (PLUGIN_DENSE_PARSER)
 
-if (PLUGIN_SYCL)
+if(PLUGIN_SYCL)
   set(CMAKE_CXX_COMPILER "icpx")
-  add_library(plugin_sycl OBJECT
-    ${xgboost_SOURCE_DIR}/plugin/sycl/data/gradient_index.cc
-    ${xgboost_SOURCE_DIR}/plugin/sycl/common/hist_util.cc
-    ${xgboost_SOURCE_DIR}/plugin/sycl/objective/regression_obj.cc
-    ${xgboost_SOURCE_DIR}/plugin/sycl/objective/multiclass_obj.cc
-    ${xgboost_SOURCE_DIR}/plugin/sycl/tree/updater_quantile_hist.cc
-    ${xgboost_SOURCE_DIR}/plugin/sycl/device_manager.cc
-    ${xgboost_SOURCE_DIR}/plugin/sycl/predictor/predictor.cc)
+  file(GLOB_RECURSE SYCL_SOURCES "sycl/*.cc")
+  add_library(plugin_sycl OBJECT ${SYCL_SOURCES})
   target_include_directories(plugin_sycl
     PRIVATE
     ${xgboost_SOURCE_DIR}/include
@@ -33,7 +27,7 @@ if (PLUGIN_SYCL)
   target_link_libraries(objxgboost PUBLIC plugin_sycl)
   # Add all objects of plugin_sycl to objxgboost
   target_sources(objxgboost INTERFACE $<TARGET_OBJECTS:plugin_sycl>)
-endif (PLUGIN_SYCL)
+endif(PLUGIN_SYCL)
 
 # Add the Federate Learning plugin if enabled.
 if (PLUGIN_FEDERATED)
